refactor(wasm): name the guest filesystem module vfs, not fs
The class is WasmVFS and the layer it adapts is runtime/vfs.py, so the
module reads vfs.py too. Each package now names its filesystem the same
way, and fs.py no longer suggests a second concept beside the core.
