+++ python/mirage/runtime/wasm/runtime.py
@@ -18,7 +18,7 @@ import threading
 from pathlib import Path
 from typing import Any
 
-from mirage.runtime.wasm.fs import WasmVFS
+from mirage.runtime.wasm.vfs import WasmVFS
 from mirage.runtime.wasm.host import WasiFs, install_wasi_fs
 
 wasmtime: Any
